Autocomplete Complete
Wrote the code to redraw the tables based on autocomplete data and pulling the info from the json, I may look into caching it in an array for greps, seems solid, but maybe unnecessary. I could host it for users after adding buttons to run the script or gather from existing data.