Trimming Branches is easy
lost a lot of time recreating content, autocomplete redraws now remove previous data as well, added search button to call the creation of the datatables, as well as allowing the dtbl to reinitialize via destroy. Srch broke for hours with no reason why and came back all of a sudden.